requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
PyMySQL==1.1.0
mysql-connector-python==8.2.0
cryptography==41.0.7
bcrypt==4.0.1
python-dotenv==1.0.0
razorpay==1.3.0
requests==2.31.0
Pillow==10.0.1
email-validator==2.1.0
